# api_gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/book/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book Related"
                ],
                "summary": "Gets details of the Book from User input book ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Book"
                        }
                    }
                }
            }
        },
...
        "/books/student/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student Book Related"
                ],
                "summary": "Retrieves the borrowed books of the student based on given ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Book"
                            }
                        }
                    }
                }
            }
        },
        "/course/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course Related"
                ],
                "summary": "Gets details of the Course from User input book ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Course"
                        }
                    }
                }
            }
        },
...
        "/courses/student/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student Course Related"
                ],
                "summary": "Retrieves the enrolled courses of the student based on given ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Course"
                            }
                        }
                    }
                }
            }
        },
        "/create/book": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book Related"
                ],
                "summary": "Creates a Book record with user input details and writes into database",
                "parameters": [
                    {
                        "description": "Book details",
                        "name": "details",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreateBookModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Book"
                        }
                    }
                }
            }
        },
        "/create/course": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course Related"
                ],
                "summary": "Creates a Course record with user input details and writes into database",
                "parameters": [
                    {
                        "description": "Course details",
                        "name": "details",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreateCourseModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Course"
                        }
                    }
                }
            }
        },
        "/create/student": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student Related"
                ],
                "summary": "Creates a student record with user input details and writes into database",
                "parameters": [
                    {
                        "description": "Student details",
                        "name": "details",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreateStudentModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Student"
                        }
                    }
                }
            }
        },
        "/delete/book/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book Related"
                ],
                "summary": "Deletes the book with the given ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Book"
                        }
                    }
                }
            }
        },
        "/delete/course/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course Related"
                ],
                "summary": "Deletes the course with the given ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Course"
                        }
                    }
                }
            }
        },
        "/delete/student/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student Related"
                ],
                "summary": "Deletes a student with the specified ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Student"
                        }
                    }
                }
            }
        },
        "/student/borrow/book": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student Book Related"
                ],
                "summary": "Borrows the book of a given Id for a given student ID",
                "parameters": [
                    {
                        "description": "Book and Student IDs",
                        "name": "student_book_ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Student_Book_Ids"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Book"
                        }
                    }
                }
            }
        },
        "/student/drop/course": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student Course Related"
                ],
                "summary": "Drop the student of the given Id from the given course(ID)",
                "parameters": [
                    {
                        "description": "Student and Course IDs",
                        "name": "student_course_ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Student_Course_Ids"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Course"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "main.Book": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "main.Course": {
            "type": "object",
            "properties": {
                "course_id": {
                    "type": "integer"
                },
                "instructor": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "main.CreateBookModel": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "main.CreateCourseModel": {
            "type": "object",
            "properties": {
                "instructor": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "main.CreateStudentModel": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                }
            }
        },
        "main.Student": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_Name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_Name": {
                    "type": "string"
                }
            }
        },
        "main.Student_Book_Ids": {
            "type": "object",
            "properties": {
                "book_id": {
                    "type": "integer"
                },
                "student_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}